requirements.txt
pandas

requests

python-dotenv

SQLAlchemy

pymysql
